extracted mappings to separate classes to improve readability
