+++ services/connectivity/mapping/src/main/java/org/eclipse/ditto/services/connectivity/mapping/javascript/MappingFunction.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2017-2018 Bosch Software Innovations GmbH.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/org/documents/epl-2.0/index.php
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.eclipse.ditto.services.connectivity.mapping.javascript;
+
+import java.util.function.Function;
+
+import org.eclipse.ditto.model.base.headers.DittoHeaders;
+import org.eclipse.ditto.model.connectivity.MessageMappingFailedException;
+import org.mozilla.javascript.RhinoException;
+
+/**
+ * Mapping interface that also provides common methods.
+ *
+ * @param <I> input data type
+ * @param <O> output data type
+ */
+public interface MappingFunction<I, O> extends Function<I, O> {
+
+    default MessageMappingFailedException buildMessageMappingFailedException(final RhinoException e,
+            final String contentType, final DittoHeaders dittoHeaders) {
+        final boolean sourceExists = e.lineSource() != null && !e.lineSource().isEmpty();
+        final String lineSource = sourceExists ? (", source:\n" + e.lineSource()) : "";
+        final boolean stackExists = e.getScriptStackTrace() != null && !e.getScriptStackTrace().isEmpty();
+        final String scriptStackTrace = stackExists ? (", stack:\n" + e.getScriptStackTrace()) : "";
+        return MessageMappingFailedException.newBuilder(contentType)
+                .description(e.getMessage() + " - in line/column #" + e.lineNumber() + "/" + e.columnNumber() +
+                        lineSource + scriptStackTrace)
+                .dittoHeaders(dittoHeaders)
+                .cause(e)
+                .build();
+    }
+}
